Replace PangeaLoadTask with a wrapper that does not need to be inserted into the pipeline.
Concept: replace the unweildy PangeaLoadTask with a set of classes
that dynamically reference the CapTask that they wrap. Currently
PLTs are inserted between normal CTs which causes problems when
calculating versions and has generally been tough to maintain.

Instead of PLTs this commit introduces PangeaCapTasks which will
still handle communication and i/o with a Pangea server but will
not need to be inserted in between pipeline tasks.

In practice PCTs are essentially subclasses of the modules they wrap
with seamless access to all methods, attributes, and classmethods of
those classes. However, this is NOT how PCTs are implemented. PCTs
define __getattr__ methods that defer to wrapped CT instances and
through metaclasses wrapped CT types. PCT types are dynamically
generated from the PCT base classes and a CT class.

The code to do this involves more python magic than is ideal. PCTs
use metaclasses as well as class factories. But I think on balance it
will be easier to maintain than PLTs.