fix: add some tests and some metaprogramming wizardry
Added some basic tests for Pangea Cap. Sadly found a tricky bug for which
there is no good resolution. I built PCTs as duck typed but luigi's
scheduler does a hard type check to see if something is a luigi task.

Making PCTs subclasses of luigi tasks causes a lot of issues, mainly
because there are lots of params involved in pangea i/o that have nothing
to do with the rest of the cap and vice versa. As such I decided it was
easier and better to spoof PCTs as luigi Tasks by modifying __class__
and __instancecheck__. This works for now but has the potential to introduce
complex problems in the future.