feat: initial approach for version reliance
In many cases we want to run a downstream task without re-running upstream tasks.
Currently this is not possible as tasks are hardcoded to look ONLY for the newest
version of the tasks it depends on. This edit will allow tasks to look for
arbitrary versions of tasks they rely on.